+++ 400-rest/src/main/java/software/wings/resources/DelegateSetupResource.java
@@ -2,6 +2,10 @@ package software.wings.resources;
 
 import static io.harness.annotations.dev.HarnessTeam.DEL;
 import static io.harness.data.structure.EmptyPredicate.isNotEmpty;
+import static io.harness.delegate.utils.RbacConstants.DELEGATE_DELETE_PERMISSION;
+import static io.harness.delegate.utils.RbacConstants.DELEGATE_EDIT_PERMISSION;
+import static io.harness.delegate.utils.RbacConstants.DELEGATE_RESOURCE_TYPE;
+import static io.harness.delegate.utils.RbacConstants.DELEGATE_VIEW_PERMISSION;
 import static io.harness.logging.AutoLogContext.OverrideBehavior.OVERRIDE_ERROR;
 
 import static software.wings.security.PermissionAttribute.PermissionType.ACCOUNT_MANAGEMENT;
@@ -16,6 +20,9 @@ import static software.wings.service.impl.DelegateServiceImpl.KUBERNETES_DELEGAT
 
 import static java.util.stream.Collectors.toList;
 
+import io.harness.accesscontrol.clients.AccessControlClient;
+import io.harness.accesscontrol.clients.Resource;
+import io.harness.accesscontrol.clients.ResourceScope;
 import io.harness.annotations.dev.HarnessModule;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.annotations.dev.TargetModule;
@@ -101,16 +108,18 @@ public class DelegateSetupResource {
   private final DelegateScopeService delegateScopeService;
   private final DownloadTokenService downloadTokenService;
   private final SubdomainUrlHelperIntfc subdomainUrlHelper;
+  private final AccessControlClient accessControlClient;
 
   @Inject
   public DelegateSetupResource(DelegateService delegateService, DelegateScopeService delegateScopeService,
       DownloadTokenService downloadTokenService, SubdomainUrlHelperIntfc subdomainUrlHelper,
-      DelegateCache delegateCache) {
+      DelegateCache delegateCache, AccessControlClient accessControlClient) {
     this.delegateService = delegateService;
     this.delegateScopeService = delegateScopeService;
     this.downloadTokenService = downloadTokenService;
     this.subdomainUrlHelper = subdomainUrlHelper;
     this.delegateCache = delegateCache;
+    this.accessControlClient = accessControlClient;
   }
 
   @GET
@@ -224,8 +233,14 @@ public class DelegateSetupResource {
   @Path("delegate-sizes")
   @Timed
   @ExceptionMetered
-  @AuthRule(permissionType = MANAGE_DELEGATES)
-  public RestResponse<List<DelegateSizeDetails>> delegateSizes(@QueryParam("accountId") @NotEmpty String accountId) {
+  @AuthRule(permissionType = LOGGED_IN)
+  // This NG specific, switching to NG access control. AuthRule to be removed also when NG access control is fully
+  // enabled.
+  public RestResponse<List<DelegateSizeDetails>> delegateSizes(@QueryParam("accountId") @NotEmpty String accountId,
+      @QueryParam("orgId") String orgId, @QueryParam("projectId") String projectId) {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(accountId, orgId, projectId),
+        Resource.of(DELEGATE_RESOURCE_TYPE, null), DELEGATE_VIEW_PERMISSION);
+
     try (AutoLogContext ignore1 = new AccountLogContext(accountId, OVERRIDE_ERROR)) {
       return new RestResponse<>(delegateService.fetchAvailableSizes());
     }
@@ -375,9 +390,15 @@ public class DelegateSetupResource {
   @Path("validate-kubernetes-yaml")
   @Timed
   @ExceptionMetered
-  @AuthRule(permissionType = MANAGE_DELEGATES)
+  @AuthRule(permissionType = LOGGED_IN)
+  // This NG specific, switching to NG access control. AuthRule to be removed also when NG access control is fully
+  // enabled.
   public RestResponse<DelegateSetupDetails> validateKubernetesYaml(@Context HttpServletRequest request,
-      @QueryParam("accountId") @NotEmpty String accountId, DelegateSetupDetails delegateSetupDetails) {
+      @QueryParam("accountId") @NotEmpty String accountId, @QueryParam("orgId") String orgId,
+      @QueryParam("projectId") String projectId, DelegateSetupDetails delegateSetupDetails) {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(accountId, orgId, projectId),
+        Resource.of(DELEGATE_RESOURCE_TYPE, null), DELEGATE_EDIT_PERMISSION);
+
     try (AutoLogContext ignore1 = new AccountLogContext(accountId, OVERRIDE_ERROR)) {
       return new RestResponse<>(delegateService.validateKubernetesYaml(accountId, delegateSetupDetails));
     }
@@ -387,10 +408,16 @@ public class DelegateSetupResource {
   @Path("generate-kubernetes-yaml")
   @Timed
   @ExceptionMetered
-  @AuthRule(permissionType = MANAGE_DELEGATES)
+  @AuthRule(permissionType = LOGGED_IN)
+  // This NG specific, switching to NG access control. AuthRule to be removed also when NG access control is fully
+  // enabled.
   public Response generateKubernetesYaml(@Context HttpServletRequest request,
-      @QueryParam("accountId") @NotEmpty String accountId, DelegateSetupDetails delegateSetupDetails,
+      @QueryParam("accountId") @NotEmpty String accountId, @QueryParam("orgId") String orgId,
+      @QueryParam("projectId") String projectId, DelegateSetupDetails delegateSetupDetails,
       @QueryParam("fileFormat") MediaType fileFormat) throws IOException {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(accountId, orgId, projectId),
+        Resource.of(DELEGATE_RESOURCE_TYPE, null), DELEGATE_EDIT_PERMISSION);
+
     try (AutoLogContext ignore1 = new AccountLogContext(accountId, OVERRIDE_ERROR)) {
       File delegateFile = delegateService.generateKubernetesYaml(accountId, delegateSetupDetails,
           subdomainUrlHelper.getManagerUrl(request, accountId), getVerificationUrl(request), fileFormat);
@@ -494,9 +521,15 @@ public class DelegateSetupResource {
   @Path("groups/{delegateGroupId}")
   @Timed
   @ExceptionMetered
-  @AuthRule(permissionType = MANAGE_DELEGATES)
+  @AuthRule(permissionType = LOGGED_IN)
+  // This NG specific, switching to NG access control. AuthRule to be removed also when NG access control is fully
+  // enabled.
   public RestResponse<Void> deleteDelegateGroup(@PathParam("delegateGroupId") @NotEmpty String delegateGroupId,
-      @QueryParam("accountId") @NotEmpty String accountId, @QueryParam("forceDelete") boolean forceDelete) {
+      @QueryParam("accountId") @NotEmpty String accountId, @QueryParam("orgId") String orgId,
+      @QueryParam("projectId") String projectId, @QueryParam("forceDelete") boolean forceDelete) {
+    accessControlClient.checkForAccessOrThrow(ResourceScope.of(accountId, orgId, projectId),
+        Resource.of(DELEGATE_RESOURCE_TYPE, delegateGroupId), DELEGATE_DELETE_PERMISSION);
+
     try (AutoLogContext ignore1 = new AccountLogContext(accountId, OVERRIDE_ERROR)) {
       delegateService.deleteDelegateGroup(accountId, delegateGroupId, forceDelete);
       return new RestResponse<>();
